Membuat class baru
Membuat class baru dengan nama Family dan Salary
