+++ Employee.java
@@ -1,105 +1,80 @@
 package lib;
 
-import java.time.LocalDate;
-import java.time.Month;
-import java.util.LinkedList;
-import java.util.List;
-
-public class Employee {
+public class Employee extends Family {
 	private enum Gender {
 		Pria,
-		Wanita
+		Wanita,
 	}
-	
+
+	private Date timeJoined;
+
 	private String employeeId;
 	private String firstName;
-	private String lastName;
-	private String idNumber;
+	private String lastName; 
 	private String address;
 
-	private Gender gender; 
+	private Gender gender;
 	private boolean isForeigner;
-	
-	private Date timeJoined;
-	
-	private int monthlySalary;
-	private int otherMonthlyIncome;
-	private int annualDeductible;
-	private int monthWorkingInYear;
-	
-	private String spouseName;
-	private String spouseIdNumber;
 
-	private List<String> childNames;
-	private List<String> childIdNumbers;
 	
-	public Employee(String employeeId, String firstName, String lastName, String idNumber, String address, Date timeJoined, boolean isForeigner, Gender gender) {
-		this.employeeId = employeeId;
-		this.firstName = firstName;
-		this.lastName = lastName;
-		this.idNumber = idNumber;
-		this.address = address;
+	public Date gettimeJoined() {
+		return timeJoined;
+	}
+
+	public void settimeJoined(Date timeJoined){
 		this.timeJoined = timeJoined;
-		this.isForeigner = isForeigner;
+	}
+
+	public String getemployeeId() {
+		return employeeId;
+	}
+
+	public void setemployeeId(String employeeId) {
+		this.employeeId = employeeId;
+	}
+
+	public String getfirstName() {
+		return firstName;
+	}
+
+	public void setfirstName(String firstName) {
+		this.firstName = firstName;
+	}
+
+	public String getlastName() {
+		return lastName;
+	}
+
+	public void setlastName(String lastName) {
+		this.lastName = lastName;
+	}
+
+	public String getaddress() {
+		return address;
+	}
+
+	public void setaddress(String address) {
+		this.address = address;
+	}
+
+	public Gender getgender() {
+		return gender;
+	}
+
+	public void setgender(Gender gender) {
 		this.gender = gender;
-		
-		childNames = new LinkedList<String>();
-		childIdNumbers = new LinkedList<String>();
+	}
+
+	public boolean getisForeigner() {
+		return isForeigner;
+	}
+
+	public void setisForeigner(boolean isForeigner) {
+		this.isForeigner = isForeigner;
 	}
 	
-	/**
-	 * Fungsi untuk menentukan gaji bulanan pegawai berdasarkan grade kepegawaiannya (grade 1: 3.000.000 per bulan, grade 2: 5.000.000 per bulan, grade 3: 7.000.000 per bulan)
-	 * Jika pegawai adalah warga negara asing gaji bulanan diperbesar sebanyak 50%
-	 */
-	
-	public void setMonthlySalary(int grade) {	
-		if (grade == 1) {
-			monthlySalary = 3000000;
-			if (isForeigner) {
-				monthlySalary = (int) (3000000 * 1.5);
-			}
-		}else if (grade == 2) {
-			monthlySalary = 5000000;
-			if (isForeigner) {
-				monthlySalary = (int) (3000000 * 1.5);
-			}
-		}else if (grade == 3) {
-			monthlySalary = 7000000;
-			if (isForeigner) {
-				monthlySalary = (int) (3000000 * 1.5);
-			}
-		}
-	}
-	
-	public void setAnnualDeductible(int deductible) {	
-		this.annualDeductible = deductible;
-	}
-	
-	public void setAdditionalIncome(int income) {	
-		this.otherMonthlyIncome = income;
-	}
-	
-	public void setSpouse(String spouseName, String spouseIdNumber) {
-		this.spouseName = spouseName;
-		this.spouseIdNumber = idNumber;
-	}
-	
-	public void addChild(String childName, String childIdNumber) {
-		childNames.add(childName);
-		childIdNumbers.add(childIdNumber);
-	}
-	
-	public int getAnnualIncomeTax() {
-		
-		//Menghitung berapa lama pegawai bekerja dalam setahun ini, jika pegawai sudah bekerja dari tahun sebelumnya maka otomatis dianggap 12 bulan.
-		LocalDate date = LocalDate.now();
-		
-		if (date.getYear() == yearJoined) {
-			monthWorkingInYear = date.getMonthValue() - monthJoined;
-		}else {
-			monthWorkingInYear = 12;
-		}
-		
-		return TaxFunction.calculateTax(monthlySalary, otherMonthlyIncome, monthWorkingInYear, annualDeductible, spouseIdNumber.equals(""), childIdNumbers.size());
-	}
 }
+
+
+
+
